Add write_text
