+++ src/fst_impls/vector_fst/vector_fst.rs
@@ -195,4 +195,4 @@ impl<W: Semiring> VectorFstState<W> {
 }
 
 add_or_fst!(W, VectorFst<W>);
-display_fst!(W, VectorFst<W>);
+display_fst_trait!(W, VectorFst<W>);
